data-collector/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
grpcio==1.60.0
grpcio-tools==1.60.0
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0
